+++ src/app/result/[id]/meal-general.tsx
@@ -1,18 +1,23 @@
 import styles from '@/styles/app/result/page.module.scss';
 import { MealRecipe } from '@/src/app/result/[id]/meal-recipe';
-import { DetailedMeal } from '@/src/types/api.types';
+import { DetailedMeal, MealRecipeSection, TranslatedIngredient } from '@/src/types/api.types';
 import { MealRating } from '@/src/app/result/[id]/meal-rating';
 import AccessTimeIcon from '@mui/icons-material/AccessTime';
 import { EditLink } from '@/src/components/common/edit-link';
 import { MealTypeText } from '@/src/types/meal.types';
 import FlatwareIcon from '@mui/icons-material/Flatware';
 import PersonAddIcon from '@mui/icons-material/PersonAdd';
+import { MealIngredients } from '@/src/app/result/[id]/meal-ingredients';
+import { MealDescription } from '@/src/app/result/[id]/meal-description';
 
 interface MealGeneralProps {
     meal: DetailedMeal;
+    description?: string;
+    ingredients?: TranslatedIngredient[];
+    recipe?: MealRecipeSection[];
 }
 
-export function MealGeneral({ meal }: MealGeneralProps) {
+export function MealGeneral({ meal, description, ingredients, recipe }: MealGeneralProps) {
     return (
         <div className={styles['result-details']}>
             <ul>
@@ -40,9 +45,10 @@ export function MealGeneral({ meal }: MealGeneralProps) {
                     {meal.provider === 'yummy' && <EditLink label={'Edytuj'} link={`/meals/create/${meal.id}`} />}
                 </li>
             </ul>
-            <div>
-                <p dangerouslySetInnerHTML={{ __html: meal.description }} />
-                <MealRecipe meal={meal} />
+            <div className={styles['meal-details-sections']}>
+                <MealDescription description={description} meal={meal} />
+                <MealIngredients ingredients={ingredients} meal={meal} />
+                <MealRecipe recipe={recipe} meal={meal} />
             </div>
         </div>
     );