+++ src/app/result/[id]/meal-ingredient-element.tsx
@@ -7,7 +7,7 @@ interface MealIngredientElementProps {
 
 export function MealIngredientElement({ text, imageUrl }: MealIngredientElementProps) {
     return (
-        <li className={`d-flex justify-content-center align-items-center ${styles['result-ingredient']}`} key={text}>
+        <li className={`d-flex justify-content-start align-items-center ${styles['result-ingredient']}`} key={text}>
             <img className={styles['result-ingredient__image']} src={imageUrl} alt={text} />
             <span className={styles['result-ingredient__text']}>{text}</span>
         </li>