+++ src/types/api.types.ts
@@ -48,6 +48,7 @@ export type MealRecipeSection = {
 
 export interface DetailedMealWithTranslations {
     meal: DetailedMeal;
+    description?: string;
     ingredients?: TranslatedIngredient[];
     recipe?: MealRecipeSection[];
 }