+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.47.0] - 2024-09-17
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Translated description of meal at result page.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Ingredients are rendered between meal description and recipe.
+
 ## [1.46.1] - 2024-09-17
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): React Strict Mode to `false` as default to prevent double renders the components on the first time.